refactor(cli): 拆分 ts_bridge transport 配置与密钥解析
- 将 resolve_transport_config_json_for 及 provider 辅助函数迁入 transport/
- Runtime 通过 TransportHost 参数化调用，不引入 trait 层
