+++ apps/cli/src/ts_bridge/mod.rs
@@ -41,6 +41,7 @@ use crate::{
 
 mod constants;
 mod json_rpc;
+mod transport;
 mod types;
 
 use constants::{ENV_API_BASE, ENV_API_KEY};
@@ -60,9 +61,9 @@ pub(crate) use types::bridge::{
 
 pub struct TsBridgeRuntime {
     process: JsonRpcProcess,
-    config: AppConfig,
-    secret_store: Arc<dyn SecretStore>,
-    workspace_root: PathBuf,
+    pub(crate) config: AppConfig,
+    pub(crate) secret_store: Arc<dyn SecretStore>,
+    pub(crate) workspace_root: PathBuf,
     session: SessionModel,
     rewind: rewind::StoredDesktopRewindMetadata,
     enabled_rules: Vec<EnabledRule>,
@@ -166,7 +167,7 @@ impl TsBridgeRuntime {
             bridge_failed: false,
             deferred_transport_replace: false,
         };
-        runtime.initialize_bridge_with_transport_config(build_mcp_only_transport_config(
+        runtime.initialize_bridge_with_transport_config(transport::build_mcp_only_transport_config(
             &runtime.workspace_root,
         ))?;
         Ok(runtime)
@@ -1402,487 +1403,14 @@ impl TsBridgeRuntime {
     }
 
     fn resolve_transport_config_json_for(&self, config: &AppConfig) -> Result<Value> {
-        let Some(active) = config
-            .active_model_profile()
-            .filter(|profile| !profile.name.trim().is_empty())
-        else {
-            return Ok(build_mcp_only_transport_config(&self.workspace_root));
-        };
-
-        let api_key = if let Ok(value) = env::var(ENV_API_KEY) {
-            let trimmed = value.trim();
-            if trimmed.is_empty() {
-                self.resolve_key_from_store(&active.name, active.provider)?
-            } else {
-                trimmed.to_string()
-            }
-        } else {
-            self.resolve_key_from_store(&active.name, active.provider)?
-        };
-
-        let api_base = env::var(ENV_API_BASE).unwrap_or_else(|_| resolve_profile_api_base(active));
-
-        let normalized_reasoning_effort = normalize_reasoning_effort_value(
-            active.reasoning_effort.clone(),
-            active.provider,
-            active.transport_kind(),
-            &active.name,
-        );
-
-        let mut transport =
-            if active.transport_kind() == crate::model_registry::ModelTransportKind::Anthropic {
-                serde_json::json!({
-                    "transportKind": "anthropic",
-                    "apiKey": api_key,
-                    "model": active.name,
-                    "baseUrl": api_base,
-                    "workspaceRoot": self.workspace_root,
-                })
-            } else if active.transport_kind() == crate::model_registry::ModelTransportKind::OpenResponses {
-                let mut transport = serde_json::json!({
-                    "transportKind": "open-responses",
-                    "apiKey": api_key,
-                    "model": active.name,
-                    "baseUrl": api_base,
-                    "workspaceRoot": self.workspace_root,
-                    "store": false,
-                });
-                if let Some(responses_provider) = open_responses_sdk_provider(active.provider) {
-                    if let Some(obj) = transport.as_object_mut() {
-                        obj.insert(
-                            "responsesProvider".to_string(),
-                            json!(responses_provider),
-                        );
-                    }
-                }
-                if active.provider == Some(ModelProvider::Azure) {
-                    let resource_name = resolve_azure_resource_name(
-                        active.azure_resource_name(),
-                        &api_base,
-                    )
-                    .ok_or_else(|| {
-                        anyhow!("Azure 模型缺少 azureResourceName 配置，请使用 Desktop 连接向导导入或 spirit model add --azure-resource-name")
-                    })?;
-                    let azure_base = azure_api_base_from_resource_name(&resource_name);
-                    if let Some(obj) = transport.as_object_mut() {
-                        obj.insert("baseUrl".to_string(), json!(azure_base));
-                        obj.insert("azureResourceName".to_string(), json!(resource_name));
-                        obj.insert("llmVendor".to_string(), json!("azure"));
-                    }
-                }
-                transport
-            } else if active.transport_kind() == crate::model_registry::ModelTransportKind::Bedrock {
-                if active.provider == Some(ModelProvider::AmazonBedrock)
-                    && crate::bedrock_mantle::is_bedrock_mantle_openai_model(&active.name)
-                {
-                    let region = active.aws_region().ok_or_else(|| {
-                        anyhow!("Amazon Bedrock 模型缺少 awsRegion 配置，请使用 Desktop 连接向导导入或手动写入 config.json")
-                    })?;
-                    let mantle_base =
-                        crate::bedrock_mantle::bedrock_mantle_api_base_from_region(&region);
-                    let mut transport = serde_json::json!({
-                        "transportKind": "open-responses",
-                        "model": active.name,
-                        "baseUrl": mantle_base,
-                        "workspaceRoot": self.workspace_root,
-                        "store": false,
-                        "responsesProvider": "openai",
-                        "llmVendor": "openai",
-                    });
-                    if !api_key.trim().is_empty() {
-                        if let Some(obj) = transport.as_object_mut() {
-                            obj.insert("apiKey".to_string(), json!(api_key));
-                        }
-                    } else if let Ok(access_key_id) =
-                        load_provider_access_key_id_from_keyring(ModelProvider::AmazonBedrock.as_str())
-                    {
-                        if let Ok(secret_access_key) = load_provider_secret_access_key_from_keyring(
-                            ModelProvider::AmazonBedrock.as_str(),
-                        ) {
-                            let access_key_id = access_key_id.trim();
-                            let secret_access_key = secret_access_key.trim();
-                            if !access_key_id.is_empty() && !secret_access_key.is_empty() {
-                                if let Some(obj) = transport.as_object_mut() {
-                                    obj.insert(
-                                        "bedrockMantleIam".to_string(),
-                                        json!({
-                                            "region": region,
-                                            "accessKeyId": access_key_id,
-                                            "secretAccessKey": secret_access_key,
-                                        }),
-                                    );
-                                }
-                            }
-                        }
-                    }
-                    transport
-                } else {
-                let region = active.aws_region().ok_or_else(|| {
-                    anyhow!("Amazon Bedrock 模型缺少 awsRegion 配置，请使用 Desktop 连接向导导入或手动写入 config.json")
-                })?;
-                let mut transport = serde_json::json!({
-                    "transportKind": "bedrock",
-                    "model": active.name,
-                    "region": region,
-                    "baseUrl": api_base,
-                    "workspaceRoot": self.workspace_root,
-                });
-                if !api_key.trim().is_empty() {
-                    if let Some(obj) = transport.as_object_mut() {
-                        obj.insert("apiKey".to_string(), json!(api_key));
-                    }
-                }
-                if let Ok(access_key_id) =
-                    load_provider_access_key_id_from_keyring(ModelProvider::AmazonBedrock.as_str())
-                {
-                    if let Ok(secret_access_key) = load_provider_secret_access_key_from_keyring(
-                        ModelProvider::AmazonBedrock.as_str(),
-                    ) {
-                        let access_key_id = access_key_id.trim();
-                        let secret_access_key = secret_access_key.trim();
-                        if !access_key_id.is_empty() && !secret_access_key.is_empty() {
-                            if let Some(obj) = transport.as_object_mut() {
-                                obj.insert("accessKeyId".to_string(), json!(access_key_id));
-                                obj.insert(
-                                    "secretAccessKey".to_string(),
-                                    json!(secret_access_key),
-                                );
-                            }
-                        }
-                    }
-                }
-                transport
-                }
-            } else {
-                serde_json::json!({
-                    "apiKey": api_key,
-                    "model": active.name,
-                    "baseUrl": api_base,
-                    "workspaceRoot": self.workspace_root,
-                })
-            };
-
-        if let Some(model_capabilities) = model_capabilities_json(active) {
-            if let Some(obj) = transport.as_object_mut() {
-                obj.insert("modelCapabilities".to_string(), model_capabilities);
-            }
-        }
-
-        if active.transport_kind() == crate::model_registry::ModelTransportKind::Anthropic {
-            if let Some(effort) = anthropic_effort_value(normalized_reasoning_effort.as_deref()) {
-                if let Some(obj) = transport.as_object_mut() {
-                    obj.insert("effort".to_string(), json!(effort));
-                }
-            }
-        } else if active.transport_kind() == crate::model_registry::ModelTransportKind::OpenResponses
-            || (active.provider == Some(ModelProvider::AmazonBedrock)
-                && crate::bedrock_mantle::is_bedrock_mantle_openai_model(&active.name))
-        {
-            let is_mantle_openai = active.provider == Some(ModelProvider::AmazonBedrock)
-                && crate::bedrock_mantle::is_bedrock_mantle_openai_model(&active.name);
-            if !is_mantle_openai {
-                if let Some(provider) = active.provider {
-                    if let Some(obj) = transport.as_object_mut() {
-                        obj.insert(
-                            "llmVendor".to_string(),
-                            json!(model_provider_vendor(provider)),
-                        );
-                    }
-                }
-            }
-            if let Some(reasoning_effort) = normalized_reasoning_effort.as_deref() {
-                if let Some(obj) = transport.as_object_mut() {
-                    obj.insert("reasoningEffort".to_string(), json!(reasoning_effort));
-                }
-            }
-        } else if active.transport_kind() == crate::model_registry::ModelTransportKind::Bedrock
-            && !(active.provider == Some(ModelProvider::AmazonBedrock)
-                && crate::bedrock_mantle::is_bedrock_mantle_openai_model(&active.name))
-        {
-            if let Some(reasoning_effort) = normalized_reasoning_effort.as_deref() {
-                if let Some(obj) = transport.as_object_mut() {
-                    obj.insert("reasoningEffort".to_string(), json!(reasoning_effort));
-                }
-            }
-        } else {
-            if let Some(provider) = active.provider {
-                if let Some(obj) = transport.as_object_mut() {
-                    obj.insert(
-                        "llmVendor".to_string(),
-                        json!(model_provider_vendor(provider)),
-                    );
-                }
-            }
-            if let Some(reasoning_effort) = normalized_reasoning_effort.as_deref() {
-                if let Some(obj) = transport.as_object_mut() {
-                    obj.insert("reasoningEffort".to_string(), json!(reasoning_effort));
-                }
-            }
-            if let Some(image_profile) = config.image_generation_model_profile() {
-                if image_profile.supports_image_generation() {
-                    if let Some(image_api_key) = self
-                        .resolve_optional_key_from_store(&image_profile.name, image_profile.provider)?
-                    {
-                        let mut image_generation = serde_json::json!({
-                            "apiKey": image_api_key,
-                            "model": image_profile.name,
-                            "baseUrl": resolve_profile_api_base(image_profile),
-                        });
-                        if let Some(provider) = image_profile.provider {
-                            if let Some(obj) = image_generation.as_object_mut() {
-                                obj.insert(
-                                    "llmVendor".to_string(),
-                                    json!(model_provider_vendor(provider)),
-                                );
-                            }
-                        }
-                        if let Some(model_capabilities) = model_capabilities_json(image_profile) {
-                            if let Some(obj) = image_generation.as_object_mut() {
-                                obj.insert("modelCapabilities".to_string(), model_capabilities);
-                            }
-                        }
-                        if let Some(obj) = transport.as_object_mut() {
-                            obj.insert("imageGeneration".to_string(), image_generation);
-                        }
-                    }
-                }
-            }
-        }
-        attach_google_vertex_transport_fields(&mut transport, active)?;
-        self.attach_video_generation_config(&mut transport, config)?;
-        Ok(transport)
-    }
-
-    fn attach_video_generation_config(
-        &self,
-        transport: &mut Value,
-        config: &AppConfig,
-    ) -> Result<()> {
-        let Some(video_profile) = config.video_generation_model_profile() else {
-            return Ok(());
-        };
-        if !video_profile.supports_video_generation() {
-            return Ok(());
-        }
-        let Some(video_api_key) = self
-            .resolve_optional_key_from_store(&video_profile.name, video_profile.provider)?
-        else {
-            return Ok(());
-        };
-
-        let mut video_generation = serde_json::json!({
-            "apiKey": video_api_key,
-            "model": video_profile.name,
-            "baseUrl": resolve_profile_api_base(video_profile),
-        });
-        if let Some(provider) = video_profile.provider {
-            if let Some(obj) = video_generation.as_object_mut() {
-                obj.insert(
-                    "llmVendor".to_string(),
-                    json!(model_provider_vendor(provider)),
-                );
-            }
-        }
-        if let Some(model_capabilities) = model_capabilities_json(video_profile) {
-            if let Some(obj) = video_generation.as_object_mut() {
-                obj.insert("modelCapabilities".to_string(), model_capabilities);
-            }
-        }
-        if let Some(obj) = transport.as_object_mut() {
-            obj.insert("videoGeneration".to_string(), video_generation);
-        }
-        Ok(())
+        transport::resolve_transport_config_json_for(
+            &transport::TransportHost::from_runtime(self),
+            config,
+        )
     }
 
     fn transport_config_will_change(&self, config: &AppConfig) -> bool {
-        if self.config.active_model != config.active_model {
-            return true;
-        }
-
-        if self
-            .config
-            .active_model_profile()
-            .map(|profile| profile.api_base.as_str())
-            != config
-                .active_model_profile()
-                .map(|profile| profile.api_base.as_str())
-        {
-            return true;
-        }
-
-        if self
-            .config
-            .active_model_profile()
-            .map(|profile| profile.transport_kind())
-            != config
-                .active_model_profile()
-                .map(|profile| profile.transport_kind())
-        {
-            return true;
-        }
-
-        if self
-            .config
-            .active_model_profile()
-            .and_then(|profile| profile.reasoning_effort.as_deref())
-            != config
-                .active_model_profile()
-                .and_then(|profile| profile.reasoning_effort.as_deref())
-        {
-            return true;
-        }
-
-        if self.config.image_generation_model != config.image_generation_model {
-            return true;
-        }
-
-        if self.config.image_generation_model_profile().map(|profile| {
-            (
-                profile.api_base.as_str(),
-                profile.provider,
-                profile.transport_kind(),
-                profile.supports_image_generation(),
-            )
-        }) != config.image_generation_model_profile().map(|profile| {
-            (
-                profile.api_base.as_str(),
-                profile.provider,
-                profile.transport_kind(),
-                profile.supports_image_generation(),
-            )
-        }) {
-            return true;
-        }
-
-        if self.config.video_generation_model != config.video_generation_model {
-            return true;
-        }
-
-        if self.config.video_generation_model_profile().map(|profile| {
-            (
-                profile.api_base.as_str(),
-                profile.provider,
-                profile.transport_kind(),
-                profile.supports_video_generation(),
-            )
-        }) != config.video_generation_model_profile().map(|profile| {
-            (
-                profile.api_base.as_str(),
-                profile.provider,
-                profile.transport_kind(),
-                profile.supports_video_generation(),
-            )
-        }) {
-            return true;
-        }
-
-        self.config
-            .active_model_profile()
-            .map(|profile| profile.provider)
-            != config
-                .active_model_profile()
-                .map(|profile| profile.provider)
-    }
-
-    fn resolve_key_from_store(
-        &self,
-        model_name: &str,
-        provider: Option<crate::model_registry::ModelProvider>,
-    ) -> Result<String> {
-        if provider == Some(ModelProvider::AmazonBedrock) {
-            if let Ok(value) =
-                crate::model_registry::load_provider_api_key_from_keyring(ModelProvider::AmazonBedrock.as_str())
-            {
-                let trimmed = value.trim();
-                if !trimmed.is_empty() {
-                    return Ok(trimmed.to_string());
-                }
-            }
-            if crate::model_registry::has_bedrock_runtime_credentials_in_keyring()? {
-                return Ok(String::new());
-            }
-        } else if provider == Some(ModelProvider::GoogleVertexAi) {
-            if let Ok(value) = crate::model_registry::load_provider_api_key_from_keyring(
-                ModelProvider::GoogleVertexAi.as_str(),
-            ) {
-                let trimmed = value.trim();
-                if !trimmed.is_empty() {
-                    return Ok(trimmed.to_string());
-                }
-            }
-            if let Some(profile) = self.config.active_model_profile() {
-                if crate::model_registry::has_google_vertex_runtime_credentials(
-                    "",
-                    profile.vertex_project().as_deref(),
-                    profile.vertex_location().as_deref(),
-                ) {
-                    return Ok(String::new());
-                }
-            }
-        } else if let Some(provider) = provider {
-            if let Ok(value) =
-                crate::model_registry::load_provider_api_key_from_keyring(provider.as_str())
-            {
-                let trimmed = value.trim();
-                if !trimmed.is_empty() {
-                    return Ok(trimmed.to_string());
-                }
-            }
-        }
-        if let Some(value) = self.secret_store.load_model_api_key(model_name)? {
-            return Ok(value);
-        }
-        if let Some(value) = self.secret_store.load_global_api_key()? {
-            return Ok(value);
-        }
-
-        if provider == Some(ModelProvider::AmazonBedrock) {
-            return Err(anyhow!(
-                "未检测到 Amazon Bedrock 凭证。请在 Desktop 连接向导配置 Bearer API Key 或 IAM 凭证，或设置环境变量 {}",
-                ENV_API_KEY
-            ));
-        }
-
-        if provider == Some(ModelProvider::GoogleVertexAi) {
-            return Err(anyhow!(
-                "未检测到 Google Vertex AI 凭证。请配置 Express API Key、服务账号（client email + private key）、或 ADC（填写 project/location 并设置 GOOGLE_APPLICATION_CREDENTIALS / gcloud 默认凭证），或设置环境变量 {}",
-                ENV_API_KEY
-            ));
-        }
-
-        Err(anyhow!(
-            "未检测到模型 {} 的 API Key。可执行 `spirit model add {} --api-base <url> --key <api_key>` 或设置环境变量 {}",
-            model_name,
-            model_name,
-            ENV_API_KEY
-        ))
-    }
-
-    fn resolve_optional_key_from_store(
-        &self,
-        model_name: &str,
-        provider: Option<crate::model_registry::ModelProvider>,
-    ) -> Result<Option<String>> {
-        if let Ok(value) = env::var(ENV_API_KEY) {
-            let trimmed = value.trim();
-            if !trimmed.is_empty() {
-                return Ok(Some(trimmed.to_string()));
-            }
-        }
-        if let Some(provider) = provider {
-            if let Ok(value) =
-                crate::model_registry::load_provider_api_key_from_keyring(provider.as_str())
-            {
-                let trimmed = value.trim();
-                if !trimmed.is_empty() {
-                    return Ok(Some(trimmed.to_string()));
-                }
-            }
-        }
-        if let Some(value) = self.secret_store.load_model_api_key(model_name)? {
-            return Ok(Some(value));
-        }
-        self.secret_store.load_global_api_key()
+        transport::transport_config_will_change(&self.config, config)
     }
 
     fn call_bridge(&mut self, method: &str, params: Option<Value>) -> Result<Value> {
@@ -2651,136 +2179,6 @@ fn approval_decision_from_input(message: &str) -> Value {
             "userMessage": message,
         }),
     }
-}
-
-fn open_responses_sdk_provider(provider: Option<ModelProvider>) -> Option<&'static str> {
-    match provider {
-        Some(ModelProvider::Openai) => Some("openai"),
-        Some(ModelProvider::Xai) => Some("xai"),
-        Some(ModelProvider::Azure) => Some("azure"),
-        Some(ModelProvider::VercelAiGateway) | Some(ModelProvider::Openrouter) => None,
-        _ => Some("open-responses-compatible"),
-    }
-}
-
-fn attach_google_vertex_transport_fields(
-    transport: &mut Value,
-    profile: &crate::model_registry::ModelProfile,
-) -> Result<()> {
-    if profile.provider != Some(ModelProvider::GoogleVertexAi) {
-        return Ok(());
-    }
-
-    let obj = transport
-        .as_object_mut()
-        .ok_or_else(|| anyhow!("transport config 不是 JSON 对象"))?;
-
-    if let Some(project) = profile.vertex_project() {
-        obj.insert("vertexProject".to_string(), json!(project));
-    }
-    if let Some(location) = profile.vertex_location() {
-        obj.insert("vertexLocation".to_string(), json!(location));
-    }
-
-    if let Ok(client_email) = crate::model_registry::load_provider_vertex_client_email_from_keyring(
-        ModelProvider::GoogleVertexAi.as_str(),
-    ) {
-        let trimmed = client_email.trim();
-        if !trimmed.is_empty() {
-            obj.insert("vertexClientEmail".to_string(), json!(trimmed));
-        }
-    }
-    if let Ok(private_key) = crate::model_registry::load_provider_vertex_private_key_from_keyring(
-        ModelProvider::GoogleVertexAi.as_str(),
-    ) {
-        let trimmed = private_key.trim();
-        if !trimmed.is_empty() {
-            obj.insert("vertexPrivateKey".to_string(), json!(trimmed));
-        }
-    }
-
-    Ok(())
-}
-
-fn model_provider_vendor(provider: ModelProvider) -> &'static str {
-    match provider {
-        ModelProvider::Deepseek => "deepseek",
-        ModelProvider::Xai => "xai",
-        ModelProvider::Moonshot => "moonshot-ai",
-        ModelProvider::ZAi => "z-ai",
-        ModelProvider::ZhipuAi => "zhipu-ai",
-        ModelProvider::Minimax => "minimax",
-        ModelProvider::Xiaomi => "xiaomi",
-        ModelProvider::Siliconflow => "siliconflow",
-        ModelProvider::Alibaba => "alibaba",
-        ModelProvider::Anthropic => {
-            unreachable!("Anthropic 不应映射到 openai-compatible llmVendor")
-        }
-        ModelProvider::VercelAiGateway => "vercel-ai-gateway",
-        ModelProvider::Openrouter => "openrouter",
-        ModelProvider::Openai => "openai",
-        ModelProvider::Google => "google",
-        ModelProvider::GoogleVertexAi => "google-vertex-ai",
-        ModelProvider::Volcengine => "volcengine",
-        ModelProvider::Azure => "azure",
-        ModelProvider::AmazonBedrock => {
-            unreachable!("Amazon Bedrock 不应映射到 openai-compatible llmVendor")
-        }
-        ModelProvider::Custom => "custom",
-    }
-}
-
-fn anthropic_effort_value(value: Option<&str>) -> Option<&'static str> {
-    match value.map(str::trim) {
-        Some("low") => Some("low"),
-        Some("medium") => Some("medium"),
-        Some("high") => Some("high"),
-        Some("xhigh") => Some("xhigh"),
-        Some("max") => Some("max"),
-        _ => None,
-    }
-}
-
-fn model_capabilities_json(profile: &crate::model_registry::ModelProfile) -> Option<Value> {
-    let capabilities = profile.explicit_capabilities()?;
-    let mut object = Map::new();
-    for capability in capabilities {
-        match capability.as_str() {
-            "chat" => {
-                object.insert("chat".to_string(), Value::Bool(true));
-            }
-            "image" | "imageInput" => {
-                object.insert("imageInput".to_string(), Value::Bool(true));
-            }
-            "audioInput" => {
-                object.insert("audioInput".to_string(), Value::Bool(true));
-            }
-            "video" | "videoInput" => {
-                object.insert("videoInput".to_string(), Value::Bool(true));
-            }
-            "imageGeneration" => {
-                object.insert("imageGeneration".to_string(), Value::Bool(true));
-            }
-            "videoGeneration" => {
-                object.insert("videoGeneration".to_string(), Value::Bool(true));
-            }
-            _ => {}
-        }
-    }
-    if object.is_empty() {
-        None
-    } else {
-        Some(Value::Object(object))
-    }
-}
-
-fn build_mcp_only_transport_config(workspace_root: &Path) -> Value {
-    json!({
-        "apiKey": "mcp-only",
-        "model": "mcp-only",
-        "baseUrl": "https://example.invalid/v1",
-        "workspaceRoot": workspace_root,
-    })
 }
 
 #[cfg(test)]
